Crate rusoto_sms

AAWS Sever Migration Service
This is the AWS Sever Migration Service API Reference. It provides descriptions, syntax, and usage examples for each of the actions and data types for the AWS Sever Migration Service (AWS SMS). The topic for each action shows the Query API request parameters and the XML response. You can also view the XML request elements in the WSDL.

Alternatively, you can use one of the AWS SDKs to access an API that's tailored to the programming language or platform that you're using. For more information, see AWS SDKs.

To learn more about the Server Migration Service, see the following resources:

AWS Sever Migration Service product page

AWS Sever Migration Service User Guide

If you're using the service, you're probably looking for ServerMigrationServiceClient and ServerMigrationService.

Structs

AppSummary
Information about the application.

Connector
Represents a connector.

CreateAppRequest

CreateAppResponse

CreateReplicationJobRequest

CreateReplicationJobResponse

DeleteAppLaunchConfigurationRequest

DeleteAppLaunchConfigurationResponse

DeleteAppReplicationConfigurationRequest

DeleteAppReplicationConfigurationResponse

DeleteAppRequest

DeleteAppResponse

DeleteReplicationJobRequest

DeleteReplicationJobResponse

DeleteServerCatalogRequest

DeleteServerCatalogResponse

DisassociateConnectorRequest

DisassociateConnectorResponse

GenerateChangeSetRequest

GenerateChangeSetResponse

GenerateTemplateRequest

GenerateTemplateResponse

GetAppLaunchConfigurationRequest

GetAppLaunchConfigurationResponse

GetAppReplicationConfigurationRequest

GetAppReplicationConfigurationResponse

GetAppRequest

GetAppResponse

GetConnectorsRequest

GetConnectorsResponse

GetReplicationJobsRequest

GetReplicationJobsResponse

GetReplicationRunsRequest

GetReplicationRunsResponse

GetServersRequest

GetServersResponse

ImportServerCatalogRequest

ImportServerCatalogResponse

LaunchAppRequest

LaunchAppResponse

LaunchDetails
Details about the latest launch of an application.

ListAppsRequest

ListAppsResponse

PutAppLaunchConfigurationRequest

PutAppLaunchConfigurationResponse

PutAppReplicationConfigurationRequest

PutAppReplicationConfigurationResponse

ReplicationJob
Represents a replication job.

ReplicationRun
Represents a replication run.

ReplicationRunStageDetails
Details of the current stage of a replication run.

S3Location
Location of the Amazon S3 object in the customer's account.

Server
Represents a server.

ServerGroup
A logical grouping of servers.

ServerGroupLaunchConfiguration
Launch configuration for a server group.

ServerGroupReplicationConfiguration
Replication configuration for a server group.

ServerLaunchConfiguration
Launch configuration for a server.

ServerMigrationServiceClient
A client for the SMS API.

ServerReplicationConfiguration
Replication configuration of a server.

ServerReplicationParameters
Replication parameters for replicating a server.

StartAppReplicationRequest

StartAppReplicationResponse

StartOnDemandReplicationRunRequest

StartOnDemandReplicationRunResponse

StopAppReplicationRequest

StopAppReplicationResponse

Tag
A label that can be assigned to an application.

TerminateAppRequest

TerminateAppResponse

UpdateAppRequest

UpdateAppResponse

UpdateReplicationJobRequest

UpdateReplicationJobResponse

UserData
A script that runs on first launch of an Amazon EC2 instance. Used for configuring the server during launch.

VmServer
Represents a VM server.

VmServerAddress
Represents a VM server location.

Enums

CreateAppError
Errors returned by CreateApp

CreateReplicationJobError
Errors returned by CreateReplicationJob

DeleteAppError
Errors returned by DeleteApp

DeleteAppLaunchConfigurationError
Errors returned by DeleteAppLaunchConfiguration

DeleteAppReplicationConfigurationError
Errors returned by DeleteAppReplicationConfiguration

DeleteReplicationJobError
Errors returned by DeleteReplicationJob

DeleteServerCatalogError
Errors returned by DeleteServerCatalog

DisassociateConnectorError
Errors returned by DisassociateConnector

GenerateChangeSetError
Errors returned by GenerateChangeSet

GenerateTemplateError
Errors returned by GenerateTemplate

GetAppError
Errors returned by GetApp

GetAppLaunchConfigurationError
Errors returned by GetAppLaunchConfiguration

GetAppReplicationConfigurationError
Errors returned by GetAppReplicationConfiguration

GetConnectorsError
Errors returned by GetConnectors

GetReplicationJobsError
Errors returned by GetReplicationJobs

GetReplicationRunsError
Errors returned by GetReplicationRuns

GetServersError
Errors returned by GetServers

ImportServerCatalogError
Errors returned by ImportServerCatalog

LaunchAppError
Errors returned by LaunchApp

ListAppsError
Errors returned by ListApps

PutAppLaunchConfigurationError
Errors returned by PutAppLaunchConfiguration

PutAppReplicationConfigurationError
Errors returned by PutAppReplicationConfiguration

StartAppReplicationError
Errors returned by StartAppReplication

StartOnDemandReplicationRunError
Errors returned by StartOnDemandReplicationRun

StopAppReplicationError
Errors returned by StopAppReplication

TerminateAppError
Errors returned by TerminateApp

UpdateAppError
Errors returned by UpdateApp

UpdateReplicationJobError
Errors returned by UpdateReplicationJob

Traits

ServerMigrationService
Trait representing the capabilities of the SMS API. SMS clients implement this trait.