Trait rusoto_sms::ServerMigrationService

pub trait ServerMigrationService {
    fn create_app(
        &self, 
        input: CreateAppRequest
    ) -> RusotoFuture<CreateAppResponse, CreateAppError>;

    fn create_replication_job(
        &self, 
        input: CreateReplicationJobRequest
    ) -> RusotoFuture<CreateReplicationJobResponse, CreateReplicationJobError>;

    fn delete_app(
        &self, 
        input: DeleteAppRequest
    ) -> RusotoFuture<DeleteAppResponse, DeleteAppError>;

    fn delete_app_launch_configuration(
        &self, 
        input: DeleteAppLaunchConfigurationRequest
    ) -> RusotoFuture<DeleteAppLaunchConfigurationResponse, DeleteAppLaunchConfigurationError>;

    fn delete_app_replication_configuration(
        &self, 
        input: DeleteAppReplicationConfigurationRequest
    ) -> RusotoFuture<DeleteAppReplicationConfigurationResponse, DeleteAppReplicationConfigurationError>;

    fn delete_replication_job(
        &self, 
        input: DeleteReplicationJobRequest
    ) -> RusotoFuture<DeleteReplicationJobResponse, DeleteReplicationJobError>;

    fn delete_server_catalog(
        &self
    ) -> RusotoFuture<DeleteServerCatalogResponse, DeleteServerCatalogError>;

    fn disassociate_connector(
        &self, 
        input: DisassociateConnectorRequest
    ) -> RusotoFuture<DisassociateConnectorResponse, DisassociateConnectorError>;

    fn generate_change_set(
        &self, 
        input: GenerateChangeSetRequest
    ) -> RusotoFuture<GenerateChangeSetResponse, GenerateChangeSetError>;

    fn generate_template(
        &self, 
        input: GenerateTemplateRequest
    ) -> RusotoFuture<GenerateTemplateResponse, GenerateTemplateError>;

    fn get_app(
        &self, 
        input: GetAppRequest
    ) -> RusotoFuture<GetAppResponse, GetAppError>;

    fn get_app_launch_configuration(
        &self, 
        input: GetAppLaunchConfigurationRequest
    ) -> RusotoFuture<GetAppLaunchConfigurationResponse, GetAppLaunchConfigurationError>;

    fn get_app_replication_configuration(
        &self, 
        input: GetAppReplicationConfigurationRequest
    ) -> RusotoFuture<GetAppReplicationConfigurationResponse, GetAppReplicationConfigurationError>;

    fn get_connectors(
        &self, 
        input: GetConnectorsRequest
    ) -> RusotoFuture<GetConnectorsResponse, GetConnectorsError>;

    fn get_replication_jobs(
        &self, 
        input: GetReplicationJobsRequest
    ) -> RusotoFuture<GetReplicationJobsResponse, GetReplicationJobsError>;

    fn get_replication_runs(
        &self, 
        input: GetReplicationRunsRequest
    ) -> RusotoFuture<GetReplicationRunsResponse, GetReplicationRunsError>;

    fn get_servers(
        &self, 
        input: GetServersRequest
    ) -> RusotoFuture<GetServersResponse, GetServersError>;

    fn import_server_catalog(
        &self
    ) -> RusotoFuture<ImportServerCatalogResponse, ImportServerCatalogError>;

    fn launch_app(
        &self, 
        input: LaunchAppRequest
    ) -> RusotoFuture<LaunchAppResponse, LaunchAppError>;

    fn list_apps(
        &self, 
        input: ListAppsRequest
    ) -> RusotoFuture<ListAppsResponse, ListAppsError>;

    fn put_app_launch_configuration(
        &self, 
        input: PutAppLaunchConfigurationRequest
    ) -> RusotoFuture<PutAppLaunchConfigurationResponse, PutAppLaunchConfigurationError>;

    fn put_app_replication_configuration(
        &self, 
        input: PutAppReplicationConfigurationRequest
    ) -> RusotoFuture<PutAppReplicationConfigurationResponse, PutAppReplicationConfigurationError>;

    fn start_app_replication(
        &self, 
        input: StartAppReplicationRequest
    ) -> RusotoFuture<StartAppReplicationResponse, StartAppReplicationError>;

    fn start_on_demand_replication_run(
        &self, 
        input: StartOnDemandReplicationRunRequest
    ) -> RusotoFuture<StartOnDemandReplicationRunResponse, StartOnDemandReplicationRunError>;

    fn stop_app_replication(
        &self, 
        input: StopAppReplicationRequest
    ) -> RusotoFuture<StopAppReplicationResponse, StopAppReplicationError>;

    fn terminate_app(
        &self, 
        input: TerminateAppRequest
    ) -> RusotoFuture<TerminateAppResponse, TerminateAppError>;

    fn update_app(
        &self, 
        input: UpdateAppRequest
    ) -> RusotoFuture<UpdateAppResponse, UpdateAppError>;

    fn update_replication_job(
        &self, 
        input: UpdateReplicationJobRequest
    ) -> RusotoFuture<UpdateReplicationJobResponse, UpdateReplicationJobError>;
}

Trait representing the capabilities of the SMS API. SMS clients implement this trait.

Required Methods

fn create_app(
&self,
input: CreateAppRequest
) -> RusotoFuture<CreateAppResponse, CreateAppError>

Creates an application. An application consists of one or more server groups. Each server group contain one or more servers.

fn create_replication_job(
&self,
input: CreateReplicationJobRequest
) -> RusotoFuture<CreateReplicationJobResponse, CreateReplicationJobError>

Creates a replication job. The replication job schedules periodic replication runs to replicate your server to AWS. Each replication run creates an Amazon Machine Image (AMI).

fn delete_app(
&self,
input: DeleteAppRequest
) -> RusotoFuture<DeleteAppResponse, DeleteAppError>

Deletes an existing application. Optionally deletes the launched stack associated with the application and all AWS SMS replication jobs for servers in the application.

fn delete_app_launch_configuration(
&self,
input: DeleteAppLaunchConfigurationRequest
) -> RusotoFuture<DeleteAppLaunchConfigurationResponse, DeleteAppLaunchConfigurationError>

Deletes existing launch configuration for an application.

fn delete_app_replication_configuration(
&self,
input: DeleteAppReplicationConfigurationRequest
) -> RusotoFuture<DeleteAppReplicationConfigurationResponse, DeleteAppReplicationConfigurationError>

Deletes existing replication configuration for an application.

fn delete_replication_job(
&self,
input: DeleteReplicationJobRequest
) -> RusotoFuture<DeleteReplicationJobResponse, DeleteReplicationJobError>

Deletes the specified replication job.

After you delete a replication job, there are no further replication runs. AWS deletes the contents of the Amazon S3 bucket used to store AWS SMS artifacts. The AMIs created by the replication runs are not deleted.

fn delete_server_catalog(
&self
) -> RusotoFuture<DeleteServerCatalogResponse, DeleteServerCatalogError>

Deletes all servers from your server catalog.

fn disassociate_connector(
&self,
input: DisassociateConnectorRequest
) -> RusotoFuture<DisassociateConnectorResponse, DisassociateConnectorError>

Disassociates the specified connector from AWS SMS.

After you disassociate a connector, it is no longer available to support replication jobs.

fn generate_change_set(
&self,
input: GenerateChangeSetRequest
) -> RusotoFuture<GenerateChangeSetResponse, GenerateChangeSetError>

Generates a target change set for a currently launched stack and writes it to an Amazon S3 object in the customer’s Amazon S3 bucket.

fn generate_template(
&self,
input: GenerateTemplateRequest
) -> RusotoFuture<GenerateTemplateResponse, GenerateTemplateError>

Generates an Amazon CloudFormation template based on the current launch configuration and writes it to an Amazon S3 object in the customer’s Amazon S3 bucket.

fn get_app(
&self,
input: GetAppRequest
) -> RusotoFuture<GetAppResponse, GetAppError>

Retrieve information about an application.

fn get_app_launch_configuration(
&self,
input: GetAppLaunchConfigurationRequest
) -> RusotoFuture<GetAppLaunchConfigurationResponse, GetAppLaunchConfigurationError>

Retrieves the application launch configuration associated with an application.

fn get_app_replication_configuration(
&self,
input: GetAppReplicationConfigurationRequest
) -> RusotoFuture<GetAppReplicationConfigurationResponse, GetAppReplicationConfigurationError>

Retrieves an application replication configuration associatd with an application.

fn get_connectors(
&self,
input: GetConnectorsRequest
) -> RusotoFuture<GetConnectorsResponse, GetConnectorsError>

Describes the connectors registered with the AWS SMS.

fn get_replication_jobs(
&self,
input: GetReplicationJobsRequest
) -> RusotoFuture<GetReplicationJobsResponse, GetReplicationJobsError>

Describes the specified replication job or all of your replication jobs.

fn get_replication_runs(
&self,
input: GetReplicationRunsRequest
) -> RusotoFuture<GetReplicationRunsResponse, GetReplicationRunsError>

Describes the replication runs for the specified replication job.

fn get_servers(
&self,
input: GetServersRequest
) -> RusotoFuture<GetServersResponse, GetServersError>

Describes the servers in your server catalog.

Before you can describe your servers, you must import them using ImportServerCatalog.

fn import_server_catalog(
&self
) -> RusotoFuture<ImportServerCatalogResponse, ImportServerCatalogError>

Gathers a complete list of on-premises servers. Connectors must be installed and monitoring all servers that you want to import.

This call returns immediately, but might take additional time to retrieve all the servers.

fn launch_app(
&self,
input: LaunchAppRequest
) -> RusotoFuture<LaunchAppResponse, LaunchAppError>

Launches an application stack.

fn list_apps(
&self,
input: ListAppsRequest
) -> RusotoFuture<ListAppsResponse, ListAppsError>

Returns a list of summaries for all applications.

fn put_app_launch_configuration(
&self,
input: PutAppLaunchConfigurationRequest
) -> RusotoFuture<PutAppLaunchConfigurationResponse, PutAppLaunchConfigurationError>

Creates a launch configuration for an application.

fn put_app_replication_configuration(
&self,
input: PutAppReplicationConfigurationRequest
) -> RusotoFuture<PutAppReplicationConfigurationResponse, PutAppReplicationConfigurationError>

Creates or updates a replication configuration for an application.

fn start_app_replication(
&self,
input: StartAppReplicationRequest
) -> RusotoFuture<StartAppReplicationResponse, StartAppReplicationError>

Starts replicating an application.

fn start_on_demand_replication_run(
&self,
input: StartOnDemandReplicationRunRequest
) -> RusotoFuture<StartOnDemandReplicationRunResponse, StartOnDemandReplicationRunError>

Starts an on-demand replication run for the specified replication job. This replication run starts immediately. This replication run is in addition to the ones already scheduled.

There is a limit on the number of on-demand replications runs you can request in a 24-hour period.

fn stop_app_replication(
&self,
input: StopAppReplicationRequest
) -> RusotoFuture<StopAppReplicationResponse, StopAppReplicationError>

Stops replicating an application.

fn terminate_app(
&self,
input: TerminateAppRequest
) -> RusotoFuture<TerminateAppResponse, TerminateAppError>

Terminates the stack for an application.

fn update_app(
&self,
input: UpdateAppRequest
) -> RusotoFuture<UpdateAppResponse, UpdateAppError>

Updates an application.

fn update_replication_job(
&self,
input: UpdateReplicationJobRequest
) -> RusotoFuture<UpdateReplicationJobResponse, UpdateReplicationJobError>

Updates the specified settings for the specified replication job.