Crate rusoto_sms

AWS Server Migration Service

If you're using the service, you're probably looking for ServerMigrationServiceClient and ServerMigrationService.

Structs

Connector
Object representing a Connector

CreateReplicationJobRequest

CreateReplicationJobResponse

DeleteReplicationJobRequest

DeleteReplicationJobResponse

DeleteServerCatalogRequest

DeleteServerCatalogResponse

DisassociateConnectorRequest

DisassociateConnectorResponse

GetConnectorsRequest

GetConnectorsResponse

GetReplicationJobsRequest

GetReplicationJobsResponse

GetReplicationRunsRequest

GetReplicationRunsResponse

GetServersRequest

GetServersResponse

ImportServerCatalogRequest

ImportServerCatalogResponse

ReplicationJob
Object representing a Replication Job

ReplicationRun
Object representing a Replication Run

Server
Object representing a server

ServerMigrationServiceClient
A client for the SMS API.

StartOnDemandReplicationRunRequest

StartOnDemandReplicationRunResponse

UpdateReplicationJobRequest

UpdateReplicationJobResponse

VmServer
Object representing a VM server

VmServerAddress
Object representing a server's location

Enums

CreateReplicationJobError
Errors returned by CreateReplicationJob

DeleteReplicationJobError
Errors returned by DeleteReplicationJob

DeleteServerCatalogError
Errors returned by DeleteServerCatalog

DisassociateConnectorError
Errors returned by DisassociateConnector

GetConnectorsError
Errors returned by GetConnectors

GetReplicationJobsError
Errors returned by GetReplicationJobs

GetReplicationRunsError
Errors returned by GetReplicationRuns

GetServersError
Errors returned by GetServers

ImportServerCatalogError
Errors returned by ImportServerCatalog

StartOnDemandReplicationRunError
Errors returned by StartOnDemandReplicationRun

UpdateReplicationJobError
Errors returned by UpdateReplicationJob

Traits

ServerMigrationService
Trait representing the capabilities of the SMS API. SMS clients implement this trait.

Type Definitions

AmiId
The AMI id for the image resulting from a Replication Run.

ConnectorCapability
Capabilities for a Connector

ConnectorCapabilityList
List of Connector Capabilities

ConnectorId
Unique Identifier for Connector

ConnectorList
List of connectors

ConnectorStatus
Status of on-premise Connector

ConnectorVersion
Connector version string

Description
The description for a Replication Job/Run.

ErrorMessage
Error Message string

Frequency
Interval between Replication Runs. This value is specified in hours, and represents the time between consecutive Replication Runs.

IpAddress
Internet Protocol (IP) Address

LicenseType
The license type to be used for the Amazon Machine Image (AMI) created after a successful ReplicationRun.

MacAddress
Hardware (MAC) address

MaxResults
The maximum number of results to return in one API call. If left empty, this will default to 50.

NextToken
Pagination token to pass as input to API call

ReplicationJobId
The unique identifier for a Replication Job.

ReplicationJobList
List of Replication Jobs

ReplicationJobState
Current state of Replication Job

ReplicationJobStatusMessage
String describing current status of Replication Job

ReplicationJobTerminated
An indicator of the Replication Job being deleted or failed.

ReplicationRunId
The unique identifier for a Replication Run.

ReplicationRunList
List of Replication Runs

ReplicationRunState
Current state of Replication Run

ReplicationRunStatusMessage
String describing current status of Replication Run

ReplicationRunType
Type of Replication Run

RoleName
Name of service role in customer's account to be used by SMS service.

ServerCatalogStatus
Status of Server catalog

ServerId
Unique Identifier for a server

ServerList
List of servers from catalog

ServerType
Type of server.

Timestamp
Timestamp of an operation

VmId
Unique Identifier for a VM

VmManagerId
Unique Identifier for VM Manager

VmManagerName
VM Manager Name

VmManagerType
VM Management Product

VmName
Name of Virtual Machine

VmPath
Path to VM