Struct rusoto_sms::ServerMigrationServiceClient

pub struct ServerMigrationServiceClient { /* fields omitted */ }

A client for the SMS API.

Methods

impl ServerMigrationServiceClient

pub fn new(region: Region) -> ServerMigrationServiceClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ServerMigrationServiceClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,